chore(jwt-handlers): promote to API v2.1
Fixes #828
