+++ src/main/java/io/cryostat/net/web/http/api/v2/AbstractJwtConsumingHandler.java
@@ -35,7 +35,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.cryostat.net.web.http.api.beta;
+package io.cryostat.net.web.http.api.v2;
 
 import java.net.MalformedURLException;
 import java.net.SocketException;
