+++ src/test/java/io/cryostat/net/web/http/api/v2/AbstractJwtConsumingHandlerTest.java
@@ -35,7 +35,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.cryostat.net.web.http.api.beta;
+package io.cryostat.net.web.http.api.v2;
 
 import java.net.URL;
 import java.net.UnknownHostException;
