+++ src/test/java/io/cryostat/net/web/http/api/v2/ReportGetHandlerTest.java
@@ -35,7 +35,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.cryostat.net.web.http.api.beta;
+package io.cryostat.net.web.http.api.v2;
 
 import java.io.File;
 import java.nio.file.Path;
@@ -45,11 +45,11 @@ import java.util.concurrent.Future;
 
 import io.cryostat.core.log.Logger;
 import io.cryostat.net.AuthManager;
+import io.cryostat.net.reports.ReportService;
 import io.cryostat.net.security.ResourceAction;
 import io.cryostat.net.security.jwt.AssetJwtHelper;
 import io.cryostat.net.web.WebServer;
 import io.cryostat.net.web.http.api.ApiVersion;
-import io.cryostat.recordings.RecordingArchiveHelper;
 import io.cryostat.recordings.RecordingNotFoundException;
 
 import com.nimbusds.jwt.JWT;
@@ -71,26 +71,26 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-class RecordingGetHandlerTest {
+class ReportGetHandlerTest {
 
-    RecordingGetHandler handler;
+    ReportGetHandler handler;
     @Mock AuthManager auth;
     @Mock AssetJwtHelper jwt;
     @Mock WebServer webServer;
-    @Mock RecordingArchiveHelper archive;
+    @Mock ReportService reports;
     @Mock Logger logger;
 
     @BeforeEach
     void setup() {
-        this.handler = new RecordingGetHandler(auth, jwt, () -> webServer, archive, logger);
+        this.handler = new ReportGetHandler(auth, jwt, () -> webServer, reports, 30, logger);
     }
 
     @Nested
     class ApiSpec {
 
         @Test
-        void shouldUseApiVersionBeta() {
-            MatcherAssert.assertThat(handler.apiVersion(), Matchers.equalTo(ApiVersion.BETA));
+        void shouldUseApiVersion2_1() {
+            MatcherAssert.assertThat(handler.apiVersion(), Matchers.equalTo(ApiVersion.V2_1));
         }
 
         @Test
@@ -101,14 +101,18 @@ class RecordingGetHandlerTest {
         @Test
         void shouldUseExpectedPath() {
             MatcherAssert.assertThat(
-                    handler.path(), Matchers.equalTo("/api/beta/recordings/:recordingName"));
+                    handler.path(), Matchers.equalTo("/api/v2.1/reports/:recordingName"));
         }
 
         @Test
         void shouldRequireResourceActions() {
             MatcherAssert.assertThat(
                     handler.resourceActions(),
-                    Matchers.equalTo(EnumSet.of(ResourceAction.READ_RECORDING)));
+                    Matchers.equalTo(
+                            EnumSet.of(
+                                    ResourceAction.READ_RECORDING,
+                                    ResourceAction.CREATE_REPORT,
+                                    ResourceAction.READ_REPORT)));
         }
 
         @Test
@@ -117,8 +121,8 @@ class RecordingGetHandlerTest {
         }
 
         @Test
-        void shouldBeOrdered() {
-            Assertions.assertTrue(handler.isOrdered());
+        void shouldNotBeOrdered() {
+            Assertions.assertFalse(handler.isOrdered());
         }
     }
 
@@ -134,7 +138,7 @@ class RecordingGetHandlerTest {
             Future<Path> future =
                     CompletableFuture.failedFuture(
                             new RecordingNotFoundException("archive", "myrecording"));
-            Mockito.when(archive.getRecordingPath(Mockito.anyString())).thenReturn(future);
+            Mockito.when(reports.get(Mockito.anyString())).thenReturn(future);
             HttpStatusException ex =
                     Assertions.assertThrows(
                             HttpStatusException.class,
@@ -154,12 +158,12 @@ class RecordingGetHandlerTest {
             Mockito.when(path.toFile()).thenReturn(file);
             Mockito.when(file.length()).thenReturn(1234L);
             Future<Path> future = CompletableFuture.completedFuture(path);
-            Mockito.when(archive.getRecordingPath(Mockito.anyString())).thenReturn(future);
+            Mockito.when(reports.get(Mockito.anyString())).thenReturn(future);
 
             handler.handleWithValidJwt(ctx, token);
 
             InOrder inOrder = Mockito.inOrder(resp);
-            inOrder.verify(resp).putHeader(HttpHeaders.CONTENT_TYPE, "application/octet-stream");
+            inOrder.verify(resp).putHeader(HttpHeaders.CONTENT_TYPE, "text/html");
             inOrder.verify(resp).putHeader(HttpHeaders.CONTENT_LENGTH, "1234");
             inOrder.verify(resp).sendFile("foo.jfr");
         }
