+++ src/test/java/io/cryostat/net/web/http/api/v2/TargetReportGetHandlerTest.java
@@ -35,10 +35,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.cryostat.net.web.http.api.beta;
+package io.cryostat.net.web.http.api.v2;
 
-import java.io.File;
-import java.nio.file.Path;
 import java.util.EnumSet;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Future;
@@ -53,6 +51,7 @@ import io.cryostat.net.web.http.api.ApiVersion;
 import io.cryostat.recordings.RecordingNotFoundException;
 
 import com.nimbusds.jwt.JWT;
+import com.nimbusds.jwt.JWTClaimsSet;
 import io.vertx.core.http.HttpHeaders;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.http.HttpServerResponse;
@@ -65,15 +64,14 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InOrder;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-class ReportGetHandlerTest {
+class TargetReportGetHandlerTest {
 
-    ReportGetHandler handler;
+    TargetReportGetHandler handler;
     @Mock AuthManager auth;
     @Mock AssetJwtHelper jwt;
     @Mock WebServer webServer;
@@ -82,15 +80,15 @@ class ReportGetHandlerTest {
 
     @BeforeEach
     void setup() {
-        this.handler = new ReportGetHandler(auth, jwt, () -> webServer, reports, 30, logger);
+        this.handler = new TargetReportGetHandler(auth, jwt, () -> webServer, reports, 30, logger);
     }
 
     @Nested
     class ApiSpec {
 
         @Test
-        void shouldUseApiVersionBeta() {
-            MatcherAssert.assertThat(handler.apiVersion(), Matchers.equalTo(ApiVersion.BETA));
+        void shouldUseApiVersion2_1() {
+            MatcherAssert.assertThat(handler.apiVersion(), Matchers.equalTo(ApiVersion.V2_1));
         }
 
         @Test
@@ -101,7 +99,8 @@ class ReportGetHandlerTest {
         @Test
         void shouldUseExpectedPath() {
             MatcherAssert.assertThat(
-                    handler.path(), Matchers.equalTo("/api/beta/reports/:recordingName"));
+                    handler.path(),
+                    Matchers.equalTo("/api/v2.1/targets/:targetId/reports/:recordingName"));
         }
 
         @Test
@@ -110,19 +109,20 @@ class ReportGetHandlerTest {
                     handler.resourceActions(),
                     Matchers.equalTo(
                             EnumSet.of(
+                                    ResourceAction.READ_TARGET,
                                     ResourceAction.READ_RECORDING,
                                     ResourceAction.CREATE_REPORT,
                                     ResourceAction.READ_REPORT)));
         }
 
         @Test
-        void shouldBeAsync() {
-            Assertions.assertTrue(handler.isAsync());
+        void shouldNotBeAsync() {
+            Assertions.assertFalse(handler.isAsync());
         }
 
         @Test
-        void shouldNotBeOrdered() {
-            Assertions.assertFalse(handler.isOrdered());
+        void shouldBeOrdered() {
+            Assertions.assertTrue(handler.isOrdered());
         }
     }
 
@@ -134,11 +134,16 @@ class ReportGetHandlerTest {
 
         @Test
         void shouldRespond404IfNotFound() throws Exception {
+            HttpServerResponse resp = Mockito.mock(HttpServerResponse.class);
+            Mockito.when(ctx.response()).thenReturn(resp);
             Mockito.when(ctx.pathParam("recordingName")).thenReturn("myrecording");
-            Future<Path> future =
+            JWTClaimsSet claims = Mockito.mock(JWTClaimsSet.class);
+            Mockito.when(claims.getStringClaim(Mockito.anyString())).thenReturn(null);
+            Mockito.when(token.getJWTClaimsSet()).thenReturn(claims);
+            Future<String> future =
                     CompletableFuture.failedFuture(
-                            new RecordingNotFoundException("archive", "myrecording"));
-            Mockito.when(reports.get(Mockito.anyString())).thenReturn(future);
+                            new RecordingNotFoundException("target", "myrecording"));
+            Mockito.when(reports.get(Mockito.any(), Mockito.anyString())).thenReturn(future);
             HttpStatusException ex =
                     Assertions.assertThrows(
                             HttpStatusException.class,
@@ -151,21 +156,16 @@ class ReportGetHandlerTest {
             HttpServerResponse resp = Mockito.mock(HttpServerResponse.class);
             Mockito.when(ctx.response()).thenReturn(resp);
             Mockito.when(ctx.pathParam("recordingName")).thenReturn("myrecording");
-            Path path = Mockito.mock(Path.class);
-            Mockito.when(path.toAbsolutePath()).thenReturn(path);
-            Mockito.when(path.toString()).thenReturn("foo.jfr");
-            File file = Mockito.mock(File.class);
-            Mockito.when(path.toFile()).thenReturn(file);
-            Mockito.when(file.length()).thenReturn(1234L);
-            Future<Path> future = CompletableFuture.completedFuture(path);
-            Mockito.when(reports.get(Mockito.anyString())).thenReturn(future);
+            JWTClaimsSet claims = Mockito.mock(JWTClaimsSet.class);
+            Mockito.when(claims.getStringClaim(Mockito.anyString())).thenReturn(null);
+            Mockito.when(token.getJWTClaimsSet()).thenReturn(claims);
+            Future<String> future = CompletableFuture.completedFuture("report text");
+            Mockito.when(reports.get(Mockito.any(), Mockito.anyString())).thenReturn(future);
 
             handler.handleWithValidJwt(ctx, token);
 
-            InOrder inOrder = Mockito.inOrder(resp);
-            inOrder.verify(resp).putHeader(HttpHeaders.CONTENT_TYPE, "text/html");
-            inOrder.verify(resp).putHeader(HttpHeaders.CONTENT_LENGTH, "1234");
-            inOrder.verify(resp).sendFile("foo.jfr");
+            Mockito.verify(resp).putHeader(HttpHeaders.CONTENT_TYPE, "text/html");
+            Mockito.verify(resp).end("report text");
         }
     }
 }
